Move all the cart-related functions from other JS files into cart.js file.
